requirements.txt
# BUFS Academic Chatbot - Dependencies
# LLM 통신
httpx>=0.24.0

# PDF 처리
PyMuPDF>=1.23.0
pdfplumber>=0.10.0

# 임베딩 & 리랭킹
sentence-transformers>=3.0.0

# 벡터 DB
# ChromaDB: Rust 백엔드(1.x)가 Windows + Python 3.12에서 쓰기 경로 segfault → pre-Rust 마지막 안정판 고정.
# chroma-hnswlib 버전은 chromadb가 자동으로 결정 (Linux/Py3.10에서는 0.7.6 사용)
chromadb>=0.6.0,<0.7.0

# 그래프
networkx>=3.0

# EN 키워드 매핑 (FlashText Aho-Corasick)
flashtext>=2.7

# BM25 sparse 검색 (하이브리드 검색: Dense + BM25 → Reranker)
rank-bm25>=0.2.2
pyyaml>=6.0

# 환경 설정
python-dotenv>=1.0.0

# UI
streamlit>=1.28.0

# 웹 크롤링
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# 첨부파일 파싱 (Office 문서)
python-docx>=1.0.0      # DOCX (Word)
openpyxl>=3.1.0         # XLSX (Excel)
olefile>=0.47           # HWP5 OLE2 파싱
xlrd>=1.2.0,<2.0.0      # XLS 레거시 BIFF (2.0+는 XLSX 불가 → 1.x 고정)

# 백그라운드 스케줄러
apscheduler>=3.10.0

# 테스트
pytest>=7.0.0
pytest-asyncio>=0.21.0

# RAGAS 평가
ragas>=0.2.6,<0.3
langchain-openai>=0.1.0
langchain-huggingface>=0.0.3
datasets>=2.14.0

# FastAPI 백엔드 (Step 1: Streamlit → FastAPI 전환)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
sse-starlette>=1.6.0

# OCR (선택 - 스캔 PDF 처리 시)
# surya-ocr
